+++ src/label.rs
@@ -11,6 +11,8 @@ use {
 pub enum Type {
     Tag,
     Ref,
+    File,
+    Dir,
 }
 
 #[derive(Clone, Debug)]
@@ -30,6 +32,8 @@ impl fmt::Display for Label {
             match self.label_type {
                 Type::Tag => "tag",
                 Type::Ref => "ref",
+                Type::File => "file",
+                Type::Dir => "dir",
             },
             self.label,
             self.path.to_string_lossy(),
@@ -38,28 +42,72 @@ impl fmt::Display for Label {
     }
 }
 
+#[derive(Clone, Debug)]
+pub struct Labels {
+    pub tags: Vec<Label>,
+    pub refs: Vec<Label>,
+    pub files: Vec<Label>,
+    pub dirs: Vec<Label>,
+}
+
 // This function returns all the labels in a file for a given type.
 pub fn parse<R: BufRead>(
     tag_regex: &Regex,
     ref_regex: &Regex,
-    label_type: Type,
+    file_regex: &Regex,
+    dir_regex: &Regex,
     path: &Path,
     reader: R,
-) -> Vec<Label> {
-    let regex = match label_type {
-        Type::Tag => tag_regex,
-        Type::Ref => ref_regex,
-    };
-
-    let mut labels: Vec<Label> = Vec::new();
+) -> Labels {
+    let mut tags: Vec<Label> = Vec::new();
+    let mut refs: Vec<Label> = Vec::new();
+    let mut files: Vec<Label> = Vec::new();
+    let mut dirs: Vec<Label> = Vec::new();
 
     for (line_number, line_result) in reader.lines().enumerate() {
         if let Ok(line) = line_result {
-            for captures in regex.captures_iter(&line) {
+            // Tags
+            for captures in tag_regex.captures_iter(&line) {
                 // If we got a match, then `captures.get(1)` is guaranteed to return a `Some`. Hence
                 // we are justified in unwrapping.
-                labels.push(Label {
-                    label_type,
+                tags.push(Label {
+                    label_type: Type::Tag,
+                    label: captures.get(1).unwrap().as_str().to_owned(),
+                    path: path.to_owned(),
+                    line_number: line_number + 1,
+                });
+            }
+
+            // Refs
+            for captures in ref_regex.captures_iter(&line) {
+                // If we got a match, then `captures.get(1)` is guaranteed to return a `Some`. Hence
+                // we are justified in unwrapping.
+                refs.push(Label {
+                    label_type: Type::Ref,
+                    label: captures.get(1).unwrap().as_str().to_owned(),
+                    path: path.to_owned(),
+                    line_number: line_number + 1,
+                });
+            }
+
+            // Files
+            for captures in file_regex.captures_iter(&line) {
+                // If we got a match, then `captures.get(1)` is guaranteed to return a `Some`. Hence
+                // we are justified in unwrapping.
+                files.push(Label {
+                    label_type: Type::File,
+                    label: captures.get(1).unwrap().as_str().to_owned(),
+                    path: path.to_owned(),
+                    line_number: line_number + 1,
+                });
+            }
+
+            // Directories
+            for captures in dir_regex.captures_iter(&line) {
+                // If we got a match, then `captures.get(1)` is guaranteed to return a `Some`. Hence
+                // we are justified in unwrapping.
+                dirs.push(Label {
+                    label_type: Type::Dir,
                     label: captures.get(1).unwrap().as_str().to_owned(),
                     path: path.to_owned(),
                     line_number: line_number + 1,
@@ -68,7 +116,12 @@ pub fn parse<R: BufRead>(
         }
     }
 
-    labels
+    Labels {
+        tags,
+        refs,
+        files,
+        dirs,
+    }
 }
 
 #[cfg(test)]
@@ -81,6 +134,8 @@ mod tests {
 
     const TAG_REGEX: &str = "(?i)\\[\\s*tag\\s*:\\s*([^\\]\\s]*)\\s*\\]";
     const REF_REGEX: &str = "(?i)\\[\\s*ref\\s*:\\s*([^\\]\\s]*)\\s*\\]";
+    const FILE_REGEX: &str = "(?i)\\[\\s*file\\s*:\\s*([^\\]\\s]*)\\s*\\]";
+    const DIR_REGEX: &str = "(?i)\\[\\s*dir\\s*:\\s*([^\\]\\s]*)\\s*\\]";
 
     #[test]
     fn parse_empty() {
@@ -89,150 +144,390 @@ mod tests {
 
         let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
         let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
 
-        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents,
+        );
 
-        assert!(tags.is_empty());
+        assert!(labels.tags.is_empty());
+        assert!(labels.refs.is_empty());
+        assert!(labels.files.is_empty());
+        assert!(labels.dirs.is_empty());
     }
 
     #[test]
     fn parse_tag_basic() {
         let path = Path::new("file.rs").to_owned();
         let contents = r"
-      [tag:label1]
+      [?tag:label]
     "
         .trim()
-        .as_bytes();
+        .replace('?', "")
+        .as_bytes()
+        .to_owned();
 
         let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
         let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
 
-        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents.as_ref(),
+        );
 
-        assert_eq!(tags.len(), 1);
-        assert_eq!(tags[0].label_type, Type::Tag);
-        assert_eq!(tags[0].label, "label1");
-        assert_eq!(tags[0].path, path);
-        assert_eq!(tags[0].line_number, 1);
+        assert_eq!(labels.tags.len(), 1);
+        assert_eq!(labels.tags[0].label_type, Type::Tag);
+        assert_eq!(labels.tags[0].label, "label");
+        assert_eq!(labels.tags[0].path, path);
+        assert_eq!(labels.tags[0].line_number, 1);
+        assert!(labels.refs.is_empty());
+        assert!(labels.files.is_empty());
+        assert!(labels.dirs.is_empty());
     }
 
     #[test]
     fn parse_ref_basic() {
         let path = Path::new("file.rs").to_owned();
         let contents = r"
-      [ref:label1]
+      [?ref:label]
     "
         .trim()
-        .as_bytes();
+        .replace('?', "")
+        .as_bytes()
+        .to_owned();
 
         let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
         let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
 
-        let refs = parse(&tag_regex, &ref_regex, Type::Ref, &path, contents);
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents.as_ref(),
+        );
 
-        assert_eq!(refs.len(), 1);
-        assert_eq!(refs[0].label_type, Type::Ref);
-        assert_eq!(refs[0].label, "label1");
-        assert_eq!(refs[0].path, path);
-        assert_eq!(refs[0].line_number, 1);
+        assert!(labels.tags.is_empty());
+        assert_eq!(labels.refs.len(), 1);
+        assert_eq!(labels.refs[0].label_type, Type::Ref);
+        assert_eq!(labels.refs[0].label, "label");
+        assert_eq!(labels.refs[0].path, path);
+        assert_eq!(labels.refs[0].line_number, 1);
+        assert!(labels.files.is_empty());
+        assert!(labels.dirs.is_empty());
     }
 
     #[test]
-    fn parse_whitespace() {
+    fn parse_file_basic() {
         let path = Path::new("file.rs").to_owned();
         let contents = r"
-      [ TAG: label2 ]
+      [?file:foo/bar/baz.txt]
     "
         .trim()
-        .as_bytes();
+        .replace('?', "")
+        .as_bytes()
+        .to_owned();
 
         let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
         let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
 
-        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents.as_ref(),
+        );
 
-        assert_eq!(tags.len(), 1);
-        assert_eq!(tags[0].label_type, Type::Tag);
-        assert_eq!(tags[0].label, "label2");
-        assert_eq!(tags[0].path, path);
-        assert_eq!(tags[0].line_number, 1);
+        assert!(labels.tags.is_empty());
+        assert!(labels.refs.is_empty());
+        assert_eq!(labels.files.len(), 1);
+        assert_eq!(labels.files[0].label_type, Type::File);
+        assert_eq!(labels.files[0].label, "foo/bar/baz.txt");
+        assert_eq!(labels.files[0].path, path);
+        assert_eq!(labels.files[0].line_number, 1);
+        assert!(labels.dirs.is_empty());
     }
 
     #[test]
-    fn parse_case() {
+    fn parse_dir_basic() {
         let path = Path::new("file.rs").to_owned();
         let contents = r"
-      [ TAG: label3 ]
-      [ TAG: LABEL3 ]
+      [?dir:foo/bar/baz]
     "
         .trim()
-        .as_bytes();
+        .replace('?', "")
+        .as_bytes()
+        .to_owned();
 
         let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
         let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
 
-        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents.as_ref(),
+        );
 
-        assert_eq!(tags.len(), 2);
-        assert_eq!(tags[0].label_type, Type::Tag);
-        assert_eq!(tags[0].label, "label3");
-        assert_eq!(tags[0].path, path);
-        assert_eq!(tags[0].line_number, 1);
-        assert_eq!(tags[1].label_type, Type::Tag);
-        assert_eq!(tags[1].label, "LABEL3");
-        assert_eq!(tags[1].path, path);
-        assert_eq!(tags[1].line_number, 2);
-        assert_ne!(tags[0].label, tags[1].label);
+        assert!(labels.tags.is_empty());
+        assert!(labels.refs.is_empty());
+        assert!(labels.files.is_empty());
+        assert_eq!(labels.dirs.len(), 1);
+        assert_eq!(labels.dirs[0].label_type, Type::Dir);
+        assert_eq!(labels.dirs[0].label, "foo/bar/baz");
+        assert_eq!(labels.dirs[0].path, path);
+        assert_eq!(labels.dirs[0].line_number, 1);
     }
 
     #[test]
     fn parse_multiple_per_line() {
         let path = Path::new("file.rs").to_owned();
         let contents = r"
-      [tag:label4][tag:label5]
+      [?tag:label][?ref:label][?file:foo/bar/baz.txt][?dir:foo/bar/baz]
     "
         .trim()
-        .as_bytes();
+        .replace('?', "")
+        .as_bytes()
+        .to_owned();
 
         let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
         let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
 
-        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents.as_ref(),
+        );
 
-        assert_eq!(tags.len(), 2);
-        assert_eq!(tags[0].label_type, Type::Tag);
-        assert_eq!(tags[0].label, "label4");
-        assert_eq!(tags[0].path, path);
-        assert_eq!(tags[0].line_number, 1);
-        assert_eq!(tags[1].label_type, Type::Tag);
-        assert_eq!(tags[1].label, "label5");
-        assert_eq!(tags[1].path, path);
-        assert_eq!(tags[1].line_number, 1);
+        assert_eq!(labels.tags.len(), 1);
+        assert_eq!(labels.tags[0].label_type, Type::Tag);
+        assert_eq!(labels.tags[0].label, "label");
+        assert_eq!(labels.tags[0].path, path);
+        assert_eq!(labels.tags[0].line_number, 1);
+
+        assert_eq!(labels.refs.len(), 1);
+        assert_eq!(labels.refs[0].label_type, Type::Ref);
+        assert_eq!(labels.refs[0].label, "label");
+        assert_eq!(labels.refs[0].path, path);
+        assert_eq!(labels.refs[0].line_number, 1);
+
+        assert_eq!(labels.files.len(), 1);
+        assert_eq!(labels.files[0].label_type, Type::File);
+        assert_eq!(labels.files[0].label, "foo/bar/baz.txt");
+        assert_eq!(labels.files[0].path, path);
+        assert_eq!(labels.files[0].line_number, 1);
+
+        assert_eq!(labels.dirs.len(), 1);
+        assert_eq!(labels.dirs[0].label_type, Type::Dir);
+        assert_eq!(labels.dirs[0].label, "foo/bar/baz");
+        assert_eq!(labels.dirs[0].path, path);
+        assert_eq!(labels.dirs[0].line_number, 1);
     }
 
     #[test]
     fn parse_multiple_lines() {
         let path = Path::new("file.rs").to_owned();
         let contents = r"
-      [tag:label6]
-      [tag:label7]
+      [?tag:label]
+      [?ref:label]
+      [?file:foo/bar/baz.txt]
+      [?dir:foo/bar/baz]
     "
         .trim()
-        .as_bytes();
+        .replace('?', "")
+        .as_bytes()
+        .to_owned();
 
         let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
         let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
 
-        let tags = parse(&tag_regex, &ref_regex, Type::Tag, &path, contents);
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents.as_ref(),
+        );
 
-        assert_eq!(tags.len(), 2);
-        assert_eq!(tags[0].label_type, Type::Tag);
-        assert_eq!(tags[0].label, "label6");
-        assert_eq!(tags[0].path, path);
-        assert_eq!(tags[0].line_number, 1);
-        assert_eq!(tags[1].label_type, Type::Tag);
-        assert_eq!(tags[1].label, "label7");
-        assert_eq!(tags[1].path, path);
-        assert_eq!(tags[1].line_number, 2);
+        assert_eq!(labels.tags.len(), 1);
+        assert_eq!(labels.tags[0].label_type, Type::Tag);
+        assert_eq!(labels.tags[0].label, "label");
+        assert_eq!(labels.tags[0].path, path);
+        assert_eq!(labels.tags[0].line_number, 1);
+
+        assert_eq!(labels.refs.len(), 1);
+        assert_eq!(labels.refs[0].label_type, Type::Ref);
+        assert_eq!(labels.refs[0].label, "label");
+        assert_eq!(labels.refs[0].path, path);
+        assert_eq!(labels.refs[0].line_number, 2);
+
+        assert_eq!(labels.files.len(), 1);
+        assert_eq!(labels.files[0].label_type, Type::File);
+        assert_eq!(labels.files[0].label, "foo/bar/baz.txt");
+        assert_eq!(labels.files[0].path, path);
+        assert_eq!(labels.files[0].line_number, 3);
+
+        assert_eq!(labels.dirs.len(), 1);
+        assert_eq!(labels.dirs[0].label_type, Type::Dir);
+        assert_eq!(labels.dirs[0].label, "foo/bar/baz");
+        assert_eq!(labels.dirs[0].path, path);
+        assert_eq!(labels.dirs[0].line_number, 4);
+    }
+
+    #[test]
+    fn parse_whitespace() {
+        let path = Path::new("file.rs").to_owned();
+        let contents = r"
+      [  ?tag   :  label            ]
+      [  ?ref   :  label            ]
+      [  ?file  :  foo/bar/baz.txt  ]
+      [  ?dir   :  foo/bar/baz      ]
+    "
+        .trim()
+        .replace('?', "")
+        .as_bytes()
+        .to_owned();
+
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
+
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents.as_ref(),
+        );
+
+        assert_eq!(labels.tags.len(), 1);
+        assert_eq!(labels.tags[0].label_type, Type::Tag);
+        assert_eq!(labels.tags[0].label, "label");
+        assert_eq!(labels.tags[0].path, path);
+        assert_eq!(labels.tags[0].line_number, 1);
+
+        assert_eq!(labels.refs.len(), 1);
+        assert_eq!(labels.refs[0].label_type, Type::Ref);
+        assert_eq!(labels.refs[0].label, "label");
+        assert_eq!(labels.refs[0].path, path);
+        assert_eq!(labels.refs[0].line_number, 2);
+
+        assert_eq!(labels.files.len(), 1);
+        assert_eq!(labels.files[0].label_type, Type::File);
+        assert_eq!(labels.files[0].label, "foo/bar/baz.txt");
+        assert_eq!(labels.files[0].path, path);
+        assert_eq!(labels.files[0].line_number, 3);
+
+        assert_eq!(labels.dirs.len(), 1);
+        assert_eq!(labels.dirs[0].label_type, Type::Dir);
+        assert_eq!(labels.dirs[0].label, "foo/bar/baz");
+        assert_eq!(labels.dirs[0].path, path);
+        assert_eq!(labels.dirs[0].line_number, 4);
+    }
+
+    #[test]
+    fn parse_case() {
+        let path = Path::new("file.rs").to_owned();
+        let contents = r"
+      [?tag:label]
+      [?TAG:LABEL]
+      [?ref:label]
+      [?REF:LABEL]
+      [?file:foo/bar/baz.txt]
+      [?FILE:FOO/BAR/BAZ.TXT]
+      [?dir:foo/bar/baz]
+      [?DIR:FOO/BAR/BAZ]
+    "
+        .trim()
+        .replace('?', "")
+        .as_bytes()
+        .to_owned();
+
+        let tag_regex: Regex = Regex::new(TAG_REGEX).unwrap();
+        let ref_regex: Regex = Regex::new(REF_REGEX).unwrap();
+        let file_regex: Regex = Regex::new(FILE_REGEX).unwrap();
+        let dir_regex: Regex = Regex::new(DIR_REGEX).unwrap();
+
+        let labels = parse(
+            &tag_regex,
+            &ref_regex,
+            &file_regex,
+            &dir_regex,
+            &path,
+            contents.as_ref(),
+        );
+
+        assert_eq!(labels.tags.len(), 2);
+        assert_eq!(labels.tags[0].label_type, Type::Tag);
+        assert_eq!(labels.tags[0].label, "label");
+        assert_eq!(labels.tags[0].path, path);
+        assert_eq!(labels.tags[0].line_number, 1);
+        assert_eq!(labels.tags[1].label_type, Type::Tag);
+        assert_eq!(labels.tags[1].label, "LABEL");
+        assert_eq!(labels.tags[1].path, path);
+        assert_eq!(labels.tags[1].line_number, 2);
+
+        assert_eq!(labels.refs.len(), 2);
+        assert_eq!(labels.refs[0].label_type, Type::Ref);
+        assert_eq!(labels.refs[0].label, "label");
+        assert_eq!(labels.refs[0].path, path);
+        assert_eq!(labels.refs[0].line_number, 3);
+        assert_eq!(labels.refs[1].label_type, Type::Ref);
+        assert_eq!(labels.refs[1].label, "LABEL");
+        assert_eq!(labels.refs[1].path, path);
+        assert_eq!(labels.refs[1].line_number, 4);
+
+        assert_eq!(labels.files.len(), 2);
+        assert_eq!(labels.files[0].label_type, Type::File);
+        assert_eq!(labels.files[0].label, "foo/bar/baz.txt");
+        assert_eq!(labels.files[0].path, path);
+        assert_eq!(labels.files[0].line_number, 5);
+        assert_eq!(labels.files[1].label_type, Type::File);
+        assert_eq!(labels.files[1].label, "FOO/BAR/BAZ.TXT");
+        assert_eq!(labels.files[1].path, path);
+        assert_eq!(labels.files[1].line_number, 6);
+
+        assert_eq!(labels.dirs.len(), 2);
+        assert_eq!(labels.dirs[0].label_type, Type::Dir);
+        assert_eq!(labels.dirs[0].label, "foo/bar/baz");
+        assert_eq!(labels.dirs[0].path, path);
+        assert_eq!(labels.dirs[0].line_number, 7);
+        assert_eq!(labels.dirs[1].label_type, Type::Dir);
+        assert_eq!(labels.dirs[1].label, "FOO/BAR/BAZ");
+        assert_eq!(labels.dirs[1].path, path);
+        assert_eq!(labels.dirs[1].line_number, 8);
     }
 }